# refactor: manually implement filter for speed